20C [0.2.1]
Menu_Version2
-Update_1
All of the buttons are arranged according to the brief.
All of them are resolution changeable ready.
The model that I used are from what we have to show what it would look like
-Update_2
Increase text size.
Change the light in the room to red instead of White and make the object affect by it